// client/src/app/api/models/bom.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Bom {
  children?: Array<Bom>;

  /**
   * the index of the article
   */
  index?: string;

  /**
   * the name of the article
   */
  name?: string;

  /**
   * the article number
   */
  number?: string;
}